Alterações refatoração: . Criação controllers de Input Jogador . Remoção de IoC e criação de factories . Versionamento de novo padrão (Controllers, Managers, Services e Entidades) . Utilização de structs no lugar de determinas classes . Remoção scripts e objetos antigos

+++ Assets/Scripts/Entities/Skills/Aggregators/BaseSkillSets.cs
@@ -4,7 +4,7 @@ namespace Assets.Scripts.Entities.Skills.Aggregators
 {
     public abstract class BaseSkillSets
     {
-        public double GetSkillValueByName(string skillName)
+        public float GetSkillValueByName(string skillName)
         {
             object skillObj = GetType().GetProperty(skillName).GetValue(this, null);
             return (float)skillObj.GetType().GetProperty("SkillValue").GetValue(skillObj, null);
